% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_mechanism.R
\docType{class}
\name{DPMechLaplace-class}
\alias{DPMechLaplace-class}
\alias{DPMechLaplace}
\alias{show,DPMechLaplace-method}
\alias{releaseResponse,DPMechLaplace,DPParamsEps-method}
\alias{sensitivityNorm,DPMechLaplace-method}
\title{An S4 class for the Laplace mechanism of differential privacy.}
\usage{
\S4method{show}{DPMechLaplace}(object)

\S4method{releaseResponse}{DPMechLaplace,DPParamsEps}(mechanism, privacyParams,
  X)

\S4method{sensitivityNorm}{DPMechLaplace}(mechanism, X1, X2)
}
\arguments{
\item{object}{an instance of class \code{DPMech}.}

\item{mechanism}{an object of class \code{\link{DPMechLaplace}}.}

\item{privacyParams}{an object of class \code{\link{DPParamsEps}}.}

\item{X}{a privacy-sensitive dataset, a list if using sensitivity sampler.}

\item{X1}{a privacy-sensitive dataset, list if sensitivity sampler compatible.}

\item{X2}{a privacy-sensitive dataset, list if sensitivity sampler compatible.}
}
\value{
list with slots per argument, actual privacy parameter; Laplace
  mechanism response with length of target release:
  \code{privacyParams, sensitivity, dim, target, response}.

scalar numeric norm of non-private \code{target} on datasets.
}
\description{
A class that implements the basic Laplace mechanism of differential privacy,
for privatizing numeric vector releases.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: automatically prints the object.

\item \code{releaseResponse}: releases Laplace mechanism responses.

\item \code{sensitivityNorm}: measures sensitivity of non-private \code{target}.
}}

\section{Slots}{

\describe{
\item{\code{sensitivity}}{non-negative scalar numeric target sensitivity.}

\item{\code{target}}{the target non-private function to be privatized.}

\item{\code{gammaSensitivity}}{\code{NA_real_} if deactive, or scalar in [0,1)
indicating that responses must be RDP with specific confidence.}

\item{\code{dim}}{positive scalar numeric dimension of responses.}
}}

\examples{
f <- function(xs) mean(xs)
n <- 100
m <- DPMechLaplace(sensitivity = 1/n, target = f, dim = 1)
X <- runif(n)
p <- DPParamsEps(epsilon = 1)
releaseResponse(m, p, X)
f <- function(xs) mean(xs)
n <- 100
m <- DPMechLaplace(sensitivity = 1/n, target = f, dim = 1)
X1 <- runif(n)
X2 <- runif(n)
sensitivityNorm(m, X1, X2)
}
\references{
Cynthia Dwork, Frank McSherry, Kobbi Nissim, and Adam Smith.
  "Calibrating noise to sensitivity in private data analysis." In Theory of
  Cryptography Conference, pp. 265-284. Springer Berlin Heidelberg, 2006.
}
