% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_params.R
\name{setDelta<-}
\alias{setDelta<-}
\title{Setter for slot \code{delta}.}
\usage{
setDelta(object) <- value
}
\arguments{
\item{object}{the instance of \code{DPParamsDel}.}

\item{value}{positive numeric \eqn{\delta} value.}
}
\description{
Use this method instead of slot \code{delta}.
}
\seealso{
\code{\link{DPParamsDel}}.
}
