% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_params.R
\name{setEpsilon<-}
\alias{setEpsilon<-}
\title{Setter for slot \code{epsilon}.}
\usage{
setEpsilon(object) <- value
}
\arguments{
\item{object}{the instance of \code{DPParamsEps}.}

\item{value}{positive numeric \eqn{\epsilon} value.}
}
\description{
Use this method instead of slot \code{epsilon}.
}
\seealso{
\code{\link{DPParamsEps}}.
}
