% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_params.R
\name{setGamma<-}
\alias{setGamma<-}
\title{Setter for slot \code{gamma}.}
\usage{
setGamma(object) <- value
}
\arguments{
\item{object}{the instance of \code{DPParamsGam}.}

\item{value}{positive numeric \eqn{\gamma} value.}
}
\description{
Use this method instead of slot \code{gamma}.
}
\seealso{
\code{\link{DPParamsGam}}.
}
