% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanisms.R
\name{releaseResponse}
\alias{releaseResponse}
\title{\code{DPMech} private release method.}
\usage{
releaseResponse(mechanism, privacyParams, X)
}
\arguments{
\item{mechanism}{an object of class \code{\link{DPMech-class}}.}

\item{privacyParams}{an object of class \code{\link{DPParamsEps}} or subclass.}

\item{X}{a privacy-sensitive dataset.}
}
\value{
list with slots per argument, including at least: actual privacy
  parameters \code{privacyParams}, and response \code{response}.
}
\description{
Runs the differentially-private mechanism on given data.
}
